Fix bug 71339 (word wrapping is applied even though it's turned off)
Don't word-wrap the quoted text when replying if word wrapping is disabled (even if smart quoting is enabled because smart quoting also strips off the signature). Eventually we should split smart quoting into 'strip off signature' and 'word-wrap quoted text' and the word-wrap column should be independent of the word-wrap option to make clear that the column is also used for word-wrapping quoted text.

svn path=/trunk/kdepim/; revision=275167