In KMFolderTree::contentsMouseReleaseEvent make sure all code paths call the base class method KFolderTree::contentsMouseReleaseEvent.
This is because KFolderTree::contentsMousePressEvent changes the
selectionMode of the listview to single, and KFolderTree assumes that
a complementary call to contentsMouseReleaseEvent is made to change
the selectionMode back to being extended.

This fixes the problem with Ctrl-Left/Right navigation of the foldertree
listview changing the selectedItem as well as the currentItem. The problem
being that after a folder was selected with the mouse the selection mode
was stuck in single mode.

svn path=/trunk/kdepim/; revision=283213