Make the reader window wait for one serial number only at any time and only update when that particular message arrives. This is necessary because when an imap message is selected the reader window gets a setMsg(0), which means it does not otherwise know it's "current" message.
Move the signal which updates from kmmainwidget to kmreaderwin where it
belongs.

This should probably be solved by not setting the msg to 0, but rather
treating it as a message with 0 body parts and then using the observer
code that is used for load on demand. I'll talk to Carsten when he gets
back from his holiday in the sun. The bastard.

CCMAIL: 62452-done@bugs.kde.org

svn path=/trunk/kdepim/; revision=300735