Add per folder option to put replies to mails in a folder into that same folder after sending instead of the indentity sent-mail folder. This overrides the identity setting, which I think makes sense. We'll have to see if it causes trouble in some way.
There is no gui for this yet, as I would appreciate the input of Aaron
"Folder Properties" Seigo as to how to add yet another option to that
dialog without making it even taller than it already is. So to try this
you'll have to add a PutRepliesInSameFolder=true to the config section
corresponding to the folder you want to set it for.

This commit message is now officially longer than the whole patch. :)

svn path=/trunk/kdepim/; revision=302497