Extension of the KDatePickerPopup.
This feature was written _before_ the feature-freeze, but got incredibly
delayed because the mailinglist did not respond. Fortunately, Reinhold
gave a green light for this feature now, thanks! I wrote in my mail that
it made the todolist's popupmenu very unstable, but somehow I've managed
to fix this (don't ask me how).

It extends the KDatePickerPopup with (optional) items like "Tomorrow",
"Next week" and "No Date".

svn path=/trunk/kdepim/; revision=316162