GUI: Merge from aegypten_branch: tie it all together: The Composer
This patch represents most of the user-visible enhancements from
aegypten_branch. It implements - among others - opportunistic encryption for
S/MIME (was OpenPGP-only), automatic choice of crypto message format (four to
choose from), per-contact crypto preferences stored in, and editable in
KAddressBook/HEAD, etc, etc.

It's a huge bulk of new code, esp. in messagecomposer, but even more so in
keyresolver (already comitted, as it's a new file), so expect a few bugs left
in there and report them (after checking for dup's, of course), to
   https://intevation.de/roundup/aegypten
(you are of course more than welcome to provide patches, now that this work is
mostly yours, anyway).

svn path=/trunk/kdepim/; revision=316515