Add presence and messaging ability to the "default table view" in KAddressbook.
To use: View->Modify View on the default table view, enable "Show instant messa
ging presence", start and connect Kopete, and watch the presence of the Kopete contacts you stored in KABC appear in the rightmost column of the view.  Right click a contact, use the File menu or toolbar items to start a chat.

svn path=/trunk/kdepim/; revision=321340