Fixed IdentityManager:  The idea of having different semantics for the const and non-const begin() and end()  doesn't work. The compiler uses the non-const one when the manager is non-const, even  if you type ConstIterator. This meant many methods were working on the shadow-identities  list instead of the real identities (e.g. the identitycombo), and it also meant  that the changed(uoid) signal was never emitted since commit() was looking into the  wrong list.
To fix this, I did like Qt: used a different name for the const and non-const
 begin/end. But I left the const ones unchanged to minimize code change, so
 I renamed the non-const begin/end to modifyBegin()/modifyEnd().

The other bug was that changed(uoid) was emitted before updating mIdentities,
so the slot (e.g. in my next kmcomposewin.cpp commit) wouldn't be able to
find that identity with identityForUoid().

CCMAIL: Marc Mutz <mutz@kde.org>

svn path=/trunk/kdepim/; revision=329612