Provide for a setting for the maximum number of simultaneous connections to the same host. This avoids the problem with some routers' firewalls that interpret lots of simultaneous connections to the same host as a SYN attack.
The maximum number of connections is configured using a configuration parameter in kmailrc called max-connections-per-host. A value of -1 (or the omission of the parameter altogether) means unlimited connections (ie same as current behvaiour). There is no means of configuring this facility via the GUI - you have to edit the kmailrc file.

svn path=/trunk/kdepim/; revision=332682