figured out the problem with duplicates when exiting and restarting akregator.
it concerned any articale that had ^M in them , since they are stripped from the description by the silplifyWhiteSpace function (from extractNode).
the problem was, archive stores article in stripped form, but we read them without stripping them. so after a restart, finding the article failed anytime.

Solution I used : I revert the default for extractNode(_,_,simplifyWhiteSpace=true) to extractNode(_,_,simplifyWhiteSpace=false) so that we don't strip for nothin, and make the stip appear once, in Article constructor

svn path=/trunk/kdepim/akregator/; revision=340047