Fix an html editing bug.
Don't show the html toolbar by default after KMail has started up, this is
actually a change in default settings. (Not sure it this should be
backported/discussed).

Remember the state of the html toolbar (as per usual in the settings, in
the case of multiple composers last closed composer wins), and restore the
html toolbar state when a new composer window is shown.

Bug fix, keep the "Formatting (HTML)" checkbox menu item in the composer
in sync with the visibility of the html toolbar.

Keep the logic that resists sending plain text emails as html. That is
even when html formatting is on and the html toolbar is shown an email
won't actually be sent as html unless an html toolbar widget is actually
used.

CCMAIL:87742-done@bugs.kde.org

svn path=/trunk/kdepim/; revision=340898