Forward port of:
CVS commit by tilladam:

Fix incidence with parents which are stored in resources by emitting
updated() when a parent is added and _not_ emitting it when a child is
added. This is necessary, since no currently used storage format writes
the list of children anyhow, only the parent in a child, so that for
resources such as the kolab one which do update via remove/re-add the
change in the parent would remove it as a parent from the children and
since the children are not remembered, restoring their parent is not
possible.

svn path=/trunk/kdepim/; revision=346169