Forward port from proko2 of:
CVS commit by tilladam:

Fix issue 139 by using the mechanism we use for setting the replied and
forwarded status after a mail has been sent succesfully to move the mail
that originally triggered the sending of the accept/decline mail, whose
serial number is remembered in a link header to the trash folder configured
for that account. In other words, if you click to accept an invitation, and
then send the generated reply, the invitation mail will automagically go
away.

svn path=/trunk/kdepim/; revision=347299