Backport of CVS commit by kloecker:
Properly clear the KMMsgPart before it's filled with the values of the DwBodyPart. This fixes bug 87198 (Sending (of new) a e-mail with kmail for kde-3.3 the name of the file attached, some times, it comes lost) which was caused by the fact that we reused the same KMMsgPart object in order to add all attachments to a forwarded message.

svn path=/branches/KDE_3_3_BRANCH/kdepim/; revision=348023