Fixed the way the "standard groupware folders" are found when using the kolab xml storage: instead of finding them by name, look for annotations like "event.default".
Fixed up mAnnotationFolderType and mAnnotationFolderTypeChanged, so that they
are only updated once we know if a folder is one of the standard groupware folders.

When the annotation isn't found and we try to create the folders, first check
for an existing folder with that name, and if it exists, change its contents type.
[This will happen to any existing proko2 user: you'll see a request for creating
the groupware folders, which will in fact only set the correct annotation on them]

Moved annotation values into kmailicalifaceimpl's array because one day Till
will need them from kmfolderimap :)

svn path=/branches/proko2/kdepim/; revision=349069