More fixes from Mario:
* Fix a bug that I detect on KOPrefsDialogColors: if a color of a category
was changed without change the combobox, the button "apply" was not enabled.

* Added mResourceColors.setAutoDelete(true)  in KOPrefs::KOPrefs to free
the colors on the map, (the colors that are assigned to the map are dinamyc
allocated ).

* The shortcuts shift+wheel and ctrl+wheel are changed to horizontal
and vertical zoom (as Reinhold Kainhofer suggested on kdepim list ).

* Don't show popup menu 'Assign Color' in resourcesview, when the item is a
subresource.

* Permit to  configurate the colors assigned to resources like the colors of
the categories is configurated. (on KOPrefsDialogColors).



The option to toggle the resource color will be added later on.

svn path=/trunk/kdepim/; revision=351126