Some long needed restructuring of the resource, namely:
o register as observer only when an incidence is really added to the cache
  calendar and actually unRegister when we don't want to listen to changes
  anymore
o don't remove and re-add incidencens on update but keep the original and
  put the corresponding uid in a list, so that the remove and add coming
  from KMail can be caught and prevented from causing actual calendar
  changes
o keep uids of incidences which are currently in transit (being processed
  by KMail) in two lists, one for ones being added, one for ones being
  deleted. This way they can be clearly identified and cleanly special
  cased. This removes quite some complexity in the add and delete handlers
o use Incidence* based add and delete methods instead of the Event/Todo/
  Journal ones because they are identical and there is no need to discern
  them
o trigger conflict resolution only when an incidence is added from KMail
  in a folder where the same uid is already present. The same uid in two
  folders is fine.


I use lists because they are expected to be very short (1 or 2 entries,
in the common case).

There are two places where resourceChanged is still emitted although it is
not necessary anymore. I will remove them as soon as I have all corner
cases with updating in KOrganizer fixed, so no regressions are introduced
until then.

I hope I haven't forgotten anything. There is some cleanup yet to come.

svn path=/branches/proko2/kdepim/; revision=352746