CVS_SILENT
Removed surplus apostrophe:

it's

only needs an apostrophe if it's short for 'it is', not if it indicates
posession.  So, for example:

It's cold today.
The dog is wearing its coat.

However, in KDE, abbreviations using apostrophes have been struck out, so
it should in fact be:

It is cold today.

svn path=/trunk/kdepim/; revision=355632